editor/requirements.txt
Flask==2.3.3
pywebview==6.0
requests==2.32.4